+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.2.x
 
 * Fixed:    Fixed missing transforms when using the MysqlCursor!
+* Fixed:    Collection::calculateAggregates() returns nulls if called on an empty collection 
 
 ### 1.2.1